requirements.txt
streamlit==1.37.1
pdfplumber==0.11.4
sentence-transformers==3.0.1
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
openai==1.37.1
tiktoken==0.7.0
pydantic==2.8.2
regex==2024.7.24
